Move TryGetFee to submodule
