+++ mempool/src/interface/mempool_interface_impl/pool/tests/expiry.rs
@@ -13,6 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common::chain::tokens::OutputValue;
+
 use super::*;
 use crate::SystemUsageEstimator;
