+++ mempool/src/interface/mempool_interface_impl/pool/tests/utils.rs
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common::chain::tokens::OutputValue;
 use common::chain::OutPoint;
 use common::primitives::H256;
